服务端升级1.5，封装EasyTcpServer,可方便创建多个服务器

+++ EasyTcpClient/EasyTcpClient.hpp
@@ -40,7 +40,7 @@ public:
 	{
 		if (INVALID_SOCKET != _sock)
 		{
-			printf("close previous connection<socket=%d>\n",_sock);
+			printf("close previous connection<socket=%d>\n", (int)_sock);
 			Close();
 		}
